// SDKs/twilio.com/twilio_preview/1.37.3/typescript/src/sdk/models/shared/previewdeployeddevicesfleetdeployment.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class PreviewDeployedDevicesFleetDeployment extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=fleet_sid" })
  fleetSid?: string;

  @SpeakeasyMetadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=sync_service_sid" })
  syncServiceSid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}